requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1
pandas==2.2.3
openpyxl==3.1.5
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
